+++ internal/blockstm/memdb_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/test-go/testify/require"
 
-	"cosmossdk.io/store/legacy/cachekv"
-	storetypes "cosmossdk.io/store/types"
+	"github.com/cosmos/cosmos-sdk/store/v2/legacy/cachekv"
+	storetypes "github.com/cosmos/cosmos-sdk/store/v2/types"
 )
 
 type (
@@ -57,7 +57,4 @@ func TestCacheWraps(t *testing.T) {
 	// attempt to cachewrap
 	cacheWrapper := storage.CacheWrap()
 	require.IsType(t, &cachekv.GStore[any]{}, cacheWrapper)
-
-	cacheWrappedWithTrace := storage.CacheWrapWithTrace(nil, nil)
-	require.IsType(t, &cachekv.GStore[any]{}, cacheWrappedWithTrace)
 }