+++ internal/blockstm/stm.go
@@ -9,8 +9,7 @@ import (
 
 	"golang.org/x/sync/errgroup"
 
-	storetypes "cosmossdk.io/store/types"
-
+	storetypes "github.com/cosmos/cosmos-sdk/store/v2/types"
 	"github.com/cosmos/cosmos-sdk/telemetry"
 )