+++ store/legacy/cachekv/store_test.go
@@ -8,9 +8,10 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"cosmossdk.io/math/unsafe"
-	"cosmossdk.io/store/dbadapter"
-	"cosmossdk.io/store/legacy/cachekv"
-	"cosmossdk.io/store/types"
+
+	"github.com/cosmos/cosmos-sdk/store/v2/dbadapter"
+	"github.com/cosmos/cosmos-sdk/store/v2/legacy/cachekv"
+	"github.com/cosmos/cosmos-sdk/store/v2/types"
 )
 
 func newCacheKVStore() types.CacheKVStore {